+++ tests/failpoints/cases/mod.rs
@@ -7,6 +7,7 @@ mod test_bootstrap;
 mod test_cmd_epoch_checker;
 mod test_conf_change;
 mod test_coprocessor;
+mod test_debugger;
 mod test_disk_full;
 mod test_early_apply;
 mod test_encryption;
